Rename shared_resources => plasma_client
This emphasizes that this is specifically for use with plasma...
