+++ oremda/__init__.py
@@ -1,5 +1,12 @@
-from .shared_resources import Client, DataArray
+from .plasma_client import PlasmaClient, DataArray
 from .operator import Operator, OperatorHandle, operator
 from .pipeline import Pipeline
 
-__all__ = ["Client", "Operator", "OperatorHandle", "Pipeline", "operator", "DataArray"]
+__all__ = [
+    "DataArray",
+    "Operator",
+    "operator",
+    "OperatorHandle",
+    "Pipeline",
+    "PlasmaClient",
+]
