+++ oremda/pipeline.py
@@ -12,7 +12,7 @@ from oremda.operator import OperatorHandle
 from oremda.utils.id import unique_id, port_id
 from oremda.typing import PortType, NodeType, IOType
 from oremda.registry import Registry
-from oremda.shared_resources import Client as MemoryClient, DataArray
+from oremda.plasma_client import DataArray, PlasmaClient
 
 
 class PipelineEdge:
@@ -120,7 +120,7 @@ def node_iter(nodes: Dict[IdType, PipelineNode], type: NodeType):
 
 class Pipeline:
     def __init__(
-        self, client: MemoryClient, registry: Registry, id: Optional[IdType] = None
+        self, client: PlasmaClient, registry: Registry, id: Optional[IdType] = None
     ):
         self._id = unique_id(id)
         self.client = client
@@ -320,7 +320,7 @@ def validate_port_type(type):
     return type
 
 
-def deserialize_pipeline(obj: JSONType, client: MemoryClient, registry: Registry):
+def deserialize_pipeline(obj: JSONType, client: PlasmaClient, registry: Registry):
     pipeline_json = PipelineJSON(**obj)
     _id = pipeline_json.id
     _nodes = pipeline_json.nodes
